Fix the issue where velocity snaps after a second
The issue was when you started moving, the velocity would be above
tangent for a second or so (sometimes longer) before snapping into the
right position.

It occurred because I was depending on the key press echo triggering
_input every frame to update the velocity, but key press echo only
starts after holding the key for a second or so.

The solution was to orient the velocity to gravity in _physics_process.