Add a basic camera controller
It's a basic third-person shooter keyboard-mouse camera controller,
which I think is a fine enough first attempt for my platforming
ambitions.

It's a little weird when the camera is near the player's +Y axis, I feel
that moving the mouse up should bring it down, but moving the mouse down
does the trick.  Might need more investigation.