Solve an issue where player walks in midair (mostly)
The issue was that the gravity center from the gravity field wasn't in
the center of the gravity field, so the trajectory was on a ~R.12 sphere
around that point instead of R.10 around the actual center.

But this points to a more important issue, namely that gravity is
presently only the notion of "down" and not a force that pushes "down".
That idea should also be implemented.